chore: Refactor ui to match figma design(issueModalTest), Add navigate function on sidebar(UserPage), Change IssueRequestPage's name, and Gather issue pages in one directory
