+++ src/App.jsx
@@ -1,10 +1,17 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import { BrowserRouter, Route, Routes, useNavigate } from 'react-router-dom'
+import { useEffect } from 'react'
+import PropTypes from 'prop-types'
+import { ToastContainer } from 'react-toastify'
+
+import brandIcon from '@assets/CoordiPAILogo.png'
+import { PageBox } from '@styles/globalStyle'
+import SideBar from '@components/SideBar'
 
 import LoginPage from './pages/login/LoginPage'
 import FirstAccountPage from './pages/AccountSetupPage/AccountSetupPage'
-import IssueModalTest from './pages/IssuePage/IssueModalTest'
+import IssueModalTest from './pages/issue/IssueModalTest'
 import UserForm from './pages/UserPage/UserPage'
-import ChangeIssuePage from './pages/IssuePage/ChangeIssuePage'
+import IssueRequestPage from './pages/issue/IssueRequestPage'
 import RepositoryCheckPage from './pages/AccountSetupPage/RepositoryCheckPage'
 import { BuildProject } from './pages/build_project/BuildProject'
 import { BuildProject2 } from './pages/build_project/BuildProject2' 
@@ -14,14 +21,6 @@ import { Home } from './pages/Home'
 import ComponentTest from '@pages/ComponentTest'
 import { SettingProject } from './pages/build_project/SettingProject'
 import IssueDetailPage from './pages/issue/IssueDetailPage'
-import PropTypes from 'prop-types'
-import { ToastContainer } from 'react-toastify'
-
-import brandIcon from '@assets/brandIcon.png'
-
-import { PageBox } from '@styles/globalStyle'
-import SideBar from '@components/SideBar'
-import { useEffect } from 'react'
 
 
 /*
@@ -39,15 +38,19 @@ function App () {
       <ToastContainer />
       <Routes>
         <Route path='/' element={<PrivateRoute element={Home} />} />
+        
         <Route path="/login" element={<LoginPage />} />
         <Route path="/firstaccount" element={<FirstAccountPage />} />
-        <Route path="/issueModalTest" element={<IssueModalTest />} />
-        <Route path="/userform" element={<UserForm />} />
+        
         <Route path="/repositorycheckpage" element={<RepositoryCheckPage />} />
         
+        {/* Sidebar */}
+        <Route path="/user" element={<UserForm />} />
+
         {/* Test Page */}
         <Route path='/components' element={<PrivateRoute element={ComponentTest} />} />
-        
+        <Route path="/issueModalTest" element={<IssueModalTest />} />
+
         {/* BuildProject Page*/}
         <Route path='/buildproject' element={<PrivateRoute element={BuildProject} />} />
         <Route path='/buildproject2' element={<PrivateRoute element={BuildProject2} />} />
@@ -57,13 +60,14 @@ function App () {
         <Route path="/project/:projectId" element={<PrivateRoute element={Project} />} />
         <Route path="/project/:projectId/issue/:issueId" element={<PrivateRoute element={IssueDetailPage} />} />
         <Route path='/project/:projectId/edit' element={<PrivateRoute element={SettingProject} />} />
-        <Route path='/project/:projectId/request/:requestId' element={<PrivateRoute element={ChangeIssuePage} />} />
+        <Route path='/project/:projectId/request/:requestId' element={<PrivateRoute element={IssueRequestPage} />} />
       </Routes>
     </BrowserRouter>
   )
 }
 
 const PrivateRoute = ({ element: Component, hasSideBar = true }) => {
+  const navigate = useNavigate()
   // const location = useLocation();
   // const { authData, isAuthorized, logout, fetchAuthData } = useAuth();
   // const { navMenus, navSubMenus, footerNavMenus } = useNav();
@@ -118,11 +122,6 @@ const PrivateRoute = ({ element: Component, hasSideBar = true }) => {
     githubId: 'testuser123'
   }
 
-  const userOnClick = () => {
-    // Handle user click event
-    console.log('User clicked!')
-  }
-
   const logout = () => {
     // Handle logout event
     console.log('Logout clicked!')
@@ -136,9 +135,10 @@ const PrivateRoute = ({ element: Component, hasSideBar = true }) => {
         <SideBar
           brandIcon={brandIcon}
           brandTitle='CoordiPai'
+          titleOnClick={() => navigate(`/`)}
           project={project}
           userInfo={userInfo}
-          userOnClick={userOnClick}
+          userOnClick={() => navigate(`/user`)}
           logout={logout}
         />
       )
